+++ packages/components/AGENTS.md
@@ -100,7 +100,7 @@ React's type definitions do not yet ship the [Invoker Commands API](https://deve
 
 Unit tests live in `configs/plugins/react/invoker-commands.spec.ts`.
 
-## Allow-discrete Ponyfill (`src/utils/allow-discrete-ponyfill.ts`)
+## Allow-discrete Ponyfill (`src/ponyfills/allow-discrete.ts`)
 
 A shared ponyfill for browsers that don't support transitioning `display` and/or `overlay` with `transition-behavior: allow-discrete` (currently Firefox and Safari). Used by the drawer to animate dialog exit transitions. See the file-level JSDoc for architecture, CSS contract, and maintenance constraints.
